+++ Kernel/Storage/IDEChannel.h
@@ -155,7 +155,7 @@ protected:
     RefPtr<StorageDevice> m_slave;
 
     RefPtr<AsyncBlockDeviceRequest> m_current_request;
-    size_t m_current_request_block_index { 0 };
+    u64 m_current_request_block_index { 0 };
     bool m_current_request_flushing_cache { false };
     SpinLock<u8> m_request_lock;
     Lock m_lock { "IDEChannel" };
